Pedantic fixes.

+++ test/nullptr_allocator_suite.cpp
@@ -20,13 +20,13 @@ TEST_F(NullPtrAllocatorTest, allocate)
 {
     Blk mem = allocator.allocate(0);
     EXPECT_EQ(mem.ptr, nullptr);
-    EXPECT_EQ(mem.size, 0);
+    EXPECT_EQ(mem.size, (uint32_t) 0);
     mem = allocator.allocate(-1);
     EXPECT_EQ(mem.ptr, nullptr);
-    EXPECT_EQ(mem.size, 0);
+    EXPECT_EQ(mem.size, (uint32_t) 0);
     mem = allocator.allocate(1);
     EXPECT_EQ(mem.ptr, nullptr);
-    EXPECT_EQ(mem.size, 0);
+    EXPECT_EQ(mem.size, (uint32_t) 0);
 }
 
 TEST_F(NullPtrAllocatorTest, deallocate_null)
@@ -34,7 +34,7 @@ TEST_F(NullPtrAllocatorTest, deallocate_null)
     Blk mem = { nullptr, 128 };
     allocator.deallocate(mem);
     EXPECT_EQ(mem.ptr, nullptr);
-    EXPECT_EQ(mem.size, 0);
+    EXPECT_EQ(mem.size, (uint32_t) 0);
 }
 
 TEST_F(NullPtrAllocatorTest, deallocate_pointer)
@@ -42,7 +42,7 @@ TEST_F(NullPtrAllocatorTest, deallocate_pointer)
     Blk mem = { reinterpret_cast<void*>(0x100), 32 };
     allocator.deallocate(mem);
     EXPECT_EQ(mem.ptr, reinterpret_cast<void*>(0x100));
-    EXPECT_EQ(mem.size, 32);
+    EXPECT_EQ(mem.size, (uint32_t) 32);
 }
 
 TEST_F(NullPtrAllocatorTest, allocate_and_deallocate)
@@ -50,7 +50,7 @@ TEST_F(NullPtrAllocatorTest, allocate_and_deallocate)
     Blk mem = allocator.allocate(16);
     allocator.deallocate(mem);
     EXPECT_EQ(mem.ptr, nullptr);
-    EXPECT_EQ(mem.size, 0);
+    EXPECT_EQ(mem.size, (uint32_t) 0);
 }
 
 TEST_F(NullPtrAllocatorTest, allocate_and_owns_and_deallocate)
@@ -59,7 +59,7 @@ TEST_F(NullPtrAllocatorTest, allocate_and_owns_and_deallocate)
     EXPECT_EQ(allocator.owns(mem), true);
     allocator.deallocate(mem);
     EXPECT_EQ(mem.ptr, nullptr);
-    EXPECT_EQ(mem.size, 0);
+    EXPECT_EQ(mem.size, (uint32_t) 0);
 }
 
 TEST_F(NullPtrAllocatorTest, owns_null_and_large_size)
